Added Scripts

+++ ECAIService/PipelineDto/ContentBasedVectorEncoded.cs
@@ -8,13 +8,13 @@ public record class ContentBasedVectorEncoded
 {
     public string VariantId { get; set; }
     public VBuffer<double> Categories { get; set; }
-    public int Weight { get; set; }
+    public double Weight { get; set; }
     public double Price { get; set; }
 
     public Vector ToVector()
     {
         var buffer = new List<float>();
-        buffer.Add(Weight);
+        buffer.Add((float)Weight);
         buffer.Add((float)Price);
         buffer.AddRange(Categories.GetValues().ToArray().Select(i => (float)i));
         return new Vector(buffer.ToArray());
